Move serialize to separate module
